# api/api-docs/build/swagger.yaml
openapi: 3.0.0
info:
  version: 1.0.0
  title: Xpense web api
  description: Xpense web api.
servers:
  - url: 'http://localhost:7000/api'
  - url: 'https://dev.api.xpense.bulldozerinc.com/api'
paths:
...
  /org:
    post:
      summary: create an organization
      tags:
        - Organization
      operationId: createOrganization
      requestBody:
        description: create organization body request
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateOrganization'
      responses:
        '201':
          description: Organization created
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '404':
          $ref: '#/components/responses/NotFound'
    get:
      summary: Get user organizations
      tags:
        - Organization
      operationId: getOrganizations
      responses:
        '200':
          description: Current user organizations
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Organizations'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '404':
          $ref: '#/components/responses/NotFound'
  '/org/{orgId}':
    get:
      summary: get organization by ID
      tags:
        - Organization
      operationId: getOrganization
      parameters:
        - in: path
          name: orgId
          schema:
            type: string
          required: true
      responses:
        '200':
          description: Current user organizations
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Organization'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '404':
          $ref: '#/components/responses/NotFound'
    put:
      summary: Update organization
      tags:
        - Organization
      parameters:
        - in: path
          name: orgId
          schema:
            type: string
          required: true
      operationId: updateOrganization
      requestBody:
        description: Update organization body request
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateOrganization'
      responses:
        '204':
          description: Organization updated.
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '404':
          $ref: '#/components/responses/NotFound'
  '/org/{orgId}/sheets':
    get:
      summary: Get user organization sheets
      tags:
        - Organization
      operationId: getSheets
      parameters:
        - in: path
          name: orgId
          schema:
            type: string
          required: true
      responses:
        '200':
          description: Current user organization sheets
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Sheet'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '404':
          $ref: '#/components/responses/NotFound'
    post:
      summary: create sheet
      tags:
        - Sheets
      parameters:
        - in: path
          name: orgId
          schema:
            type: string
          required: true
      operationId: createSheet
      requestBody:
        description: create sheet body request
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Sheet'
      responses:
        '201':
          description: Sheet created
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '404':
          $ref: '#/components/responses/NotFound'
  '/org/{orgId}/sheets/{sheetId}':
    get:
      summary: get sheet by id
      tags:
        - Organization
        - Sheet
      operationId: getSheetById
      parameters:
        - in: path
          name: orgId
          schema:
            type: string
          required: true
        - in: path
          name: sheetId
          schema:
            type: string
          required: true
      responses:
        '200':
          description: Get sheet by id
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Sheet'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      summary: delete sheet
      tags:
        - Organization
        - Sheet
      operationId: deleteSheet
      parameters:
        - in: path
          name: orgId
          schema:
            type: string
          required: true
        - in: path
          name: sheetId
          schema:
            type: string
          required: true
      responses:
        '204':
          description: Sheet successfully delete.
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '404':
          $ref: '#/components/responses/NotFound'
  '/org/{orgId}/groups':
    get:
      summary: Get groups by org
      tags:
        - Organization
        - Group
      operationId: groupsList
      parameters:
        - in: path
          name: orgId
          schema:
            type: string
          required: true
      responses:
        '200':
          description: Group created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Groups'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
    post:
      summary: Create new group
      tags:
        - Organization
        - Group
      operationId: createGroup
      parameters:
        - in: path
          name: orgId
          schema:
            type: string
          required: true
      requestBody:
        description: Group DTO object
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateGroup'
      responses:
        '201':
          description: Group created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Group'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '404':
          $ref: '#/components/responses/NotFound'
  '/org/{orgId}/groups/{groupId}':
    get:
      summary: get group by id
      tags:
        - Organization
        - Group
      operationId: getGroupById
      parameters:
        - in: path
          name: orgId
          schema:
            type: string
          required: true
        - in: path
          name: groupId
          schema:
            type: string
          required: true
      responses:
        '200':
          description: Get group by id
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Group'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '404':
          $ref: '#/components/responses/NotFound'
    put:
      summary: Update group
      tags:
        - Organization
        - Group
      operationId: updateGroup
      parameters:
        - in: path
          name: orgId
          schema:
            type: string
          required: true
        - in: path
          name: groupId
          schema:
            type: string
          required: true
      requestBody:
        description: Group DTO object
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateGroup'
      responses:
        '204':
          description: Role successfully updated.
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      summary: delete group
      tags:
        - Organization
        - Group
      operationId: deleteGroup
      parameters:
        - in: path
          name: orgId
          schema:
            type: string
          required: true
        - in: path
          name: groupId
          schema:
            type: string
          required: true
      responses:
        '204':
          description: Role successfully updated.
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '404':
          $ref: '#/components/responses/NotFound'
  '/org/{orgId}/groups/{groupId}/members':
    patch:
      summary: Assign user to a group
      tags:
        - Organization
        - Group
      operationId: assingUserToGroup
      parameters:
        - in: path
          name: orgId
          schema:
            type: string
          required: true
        - in: path
          name: groupId
          schema:
            type: string
          required: true
      requestBody:
        description: User ID
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                userId:
                  type: string
      responses:
        '204':
          description: User successfully assigned to the group.
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '404':
          $ref: '#/components/responses/NotFound'
  '/org/{orgId}/groups/{groupId}/members/{userId}':
    delete:
      summary: Remove user from group
      tags:
        - Organization
        - Group
      operationId: removeUserFromGroup
      parameters:
        - in: path
          name: orgId
          schema:
            type: string
          required: true
        - in: path
          name: groupId
          schema:
            type: string
          required: true
        - in: path
          name: userId
          schema:
            type: string
          required: true
      responses:
        '204':
          description: User successfully removed from the group.
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '404':
          $ref: '#/components/responses/NotFound'
  '/org/{orgId}/groups/{groupId}/expense-types':
    patch:
      summary: Assign expense types to a group
      tags:
        - Organization
        - Group
      operationId: assingExpenseTypesToGroup
      parameters:
        - in: path
          name: orgId
          schema:
            type: string
          required: true
        - in: path
          name: groupId
          schema:
            type: string
          required: true
      requestBody:
        description: Expense types IDS
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                expenseTypes:
                  type: array
                  items:
                    type: string
      responses:
        '204':
          description: Expense types successfully assigned to the group.
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '404':
          $ref: '#/components/responses/NotFound'
...
  '/org/{orgId}/exp-types':
    get:
      summary: Get list of expense types
      tags:
        - Expense types
        - Organization
      operationId: getExpenseTypes
      parameters:
        - in: path
          name: orgId
          schema:
            type: string
          required: true
      responses:
        '200':
          description: List of expense types
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExpenseTypes'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '404':
          $ref: '#/components/responses/NotFound'
    post:
      summary: Create new expense type
      tags:
        - Organization
        - Expense type
      operationId: createExpenseType
      parameters:
        - in: path
          name: orgId
          schema:
            type: string
          required: true
      requestBody:
        description: Expense type DTO object
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateExpenseType'
      responses:
        '201':
          description: Expense type created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExpenseType'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '404':
          $ref: '#/components/responses/NotFound'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Organization:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        status:
          type: string
        contact:
          type: object
          properties:
            name:
              type: string
            phone:
              type: string
            email:
              type: string
    CreateOrganization:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        status:
          type: string
        contact:
          type: object
          properties:
            name:
              type: string
            phone:
              type: string
            email:
              type: string
    Organizations:
      type: array
      items:
        $ref: '#/components/schemas/Organization'
    UpdateOrganization:
      type: object
      properties:
        name:
          type: string
          nullable: true
        contact:
          type: object
          properties:
            name:
              type: string
              nullable: true
            phone:
              type: string
              nullable: true
            email:
              type: string
              nullable: true
    Sheet:
      type: object
      properties:
        id:
          type: string
        label:
          type: string
        status:
          type: string
        creationDate:
          type: string
          format: date-time
        submittedOn:
          type: string
          format: date-time
        exported:
          type: boolean
        total:
          type: integer
...
    Group:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        orgId:
          type: string
        members:
          type: array
          items:
            type: string
        expenseTypes:
          type: array
          items:
            type: string
    Groups:
      type: array
      items:
        $ref: '#/components/schemas/Group'
    CreateGroup:
      type: object
      properties:
        name:
          type: string
...
    ExpenseTypes:
      type: array
      items:
        $ref: '#/components/schemas/ExpenseType'
    ExpenseType:
      type: object
      properties:
        id:
          type: string
        code:
          type: string
        type:
          type: string
        multiplicator:
          type: integer
          example: 0.25
        status:
          type: string
        i18n:
          type: string
        attachmentRequired:
          type: boolean
    CreateExpenseType:
      type: object
      properties:
        code:
          type: string
        type:
          type: string
        multiplicator:
          type: integer
          example: 0.25
        status:
          type: string
        i18n:
          type: string
        attachmentRequired:
          type: boolean
...
  responses:
    UnauthorizedError:
      description: Access token is missing or invalid
    NotFound:
      description: The specified resource was not found
    BadRequest:
      description: Bad request
security:
  - bearerAuth: []